+++ Source/SwiftLintBuiltInRules/Rules/Style/PreferSelfInStaticReferencesRuleExamples.swift
@@ -113,6 +113,50 @@ enum PreferSelfInStaticReferencesRuleExamples {
                 struct S1 {}
             }
             """, excludeFromDocumentation: true),
+        Example("""
+            class Foo {
+                static let i = 0
+            }
+            class Bar {}
+            extension Bar {
+                func f() -> Int { Foo.i }
+            }
+            """, excludeFromDocumentation: true),
+        Example("""
+            class Outer {
+                class Inner {
+                    static let i = 0
+                }
+                class Other {
+                    static let i = 0
+                }
+            }
+            extension Outer.Inner {
+                func f() -> Int { Outer.Other.i }
+            }
+            """, excludeFromDocumentation: true),
+        Example("""
+            enum Outer {
+                static let i = 0
+                struct Inner {
+                    static let j = 0
+                }
+            }
+            extension Outer.Inner {
+                func f() -> Int { Outer.i }
+            }
+            """, excludeFromDocumentation: true),
+        Example("""
+            class Outer {
+                class Inner<T> {
+                    static let i = 0
+                }
+            }
+            extension Outer.Inner {
+                func f() { _ = Outer.Inner<Int>() }
+                func g() { _ = Outer.Inner<Int>.i }
+            }
+            """, excludeFromDocumentation: true),
     ]
 
     static let triggeringExamples = [
@@ -235,6 +279,37 @@ enum PreferSelfInStaticReferencesRuleExamples {
                 }
             }
             """, excludeFromDocumentation: true),
+        Example("""
+            class C {
+                static let i = 0
+            }
+            extension C {
+                func f() -> Int { ↓C.i }
+                var v: Int { ↓C.i }
+            }
+            """, excludeFromDocumentation: true),
+        Example("""
+            class Outer {
+                class Inner {
+                    static let i = 0
+                }
+            }
+            extension Outer.Inner {
+                func f() -> Int { ↓Outer.Inner.i }
+            }
+            """, excludeFromDocumentation: true),
+        Example("""
+            enum Outer {
+                enum Middle {
+                    struct Inner {
+                        static let i = 0
+                    }
+                }
+            }
+            extension Outer.Middle.Inner {
+                func f() -> Int { ↓Outer.Middle.Inner.i }
+            }
+            """, excludeFromDocumentation: true),
     ]
 
     static let corrections = [
@@ -275,6 +350,40 @@ enum PreferSelfInStaticReferencesRuleExamples {
                     let k = Self  . j
                     static func f(_ l: Int = Self.i) -> Int { l*Self.j }
                     func g() { Self.i + Self.f() + k }
+                }
+                """),
+        Example("""
+            class C {
+                static let i = 0
+            }
+            extension C {
+                func f() -> Int { ↓C.i }
+            }
+            """): Example("""
+                class C {
+                    static let i = 0
+                }
+                extension C {
+                    func f() -> Int { Self.i }
+                }
+                """),
+        Example("""
+            class Outer {
+                class Inner {
+                    static let i = 0
+                }
+            }
+            extension Outer.Inner {
+                func f() -> Int { ↓Outer.Inner.i }
+            }
+            """): Example("""
+                class Outer {
+                    class Inner {
+                        static let i = 0
+                    }
+                }
+                extension Outer.Inner {
+                    func f() -> Int { Self.i }
                 }
                 """),
     ]